+++ packages/vault/test/mocks/mock_restorable.dart
@@ -1,0 +1,4 @@
+import 'package:affinidi_tdk_vault/affinidi_tdk_vault.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockRestorable extends Mock implements Restorable {}