+++ packages/vault_data_manager/test/mocks/mock_cryptography_setup.dart
@@ -6,13 +6,6 @@ import 'mock_cryptography_service.dart';
 void setupMockCryptographyDefaults(MockCryptographyService mock) {
   when(() => mock.getRandomBytes(32)).thenReturn(mockBytes());
 
-  when(
-    () => mock.Pbkdf2(
-      password: any(named: 'password'),
-      nonce: any(named: 'nonce'),
-    ),
-  ).thenAnswer((_) async => mockBytes());
-
   when(() => mock.encryptToBytes(any(), any())).thenReturn(mockBytes());
 
   when(() => mock.decryptFromBytes(any(), any())).thenReturn(mockBytes());